adds value events

+++ crates/engine/src/metronome.rs
@@ -121,7 +121,7 @@ impl Metronome {
                 let mut beat_index = ceil_nonneg(block.p0 / step) as i64;
                 let mut position = beat_index as f64 * step;
                 while position < block.p1 {
-                    let offset = floor_nonneg(pulses_to_samples(position - block.p0, block.bpm, self.sample_rate as f64)) as usize;
+                    let offset = floor_nonneg(pulses_to_samples(position - block.p0, block.bpm, self.sample_rate)) as usize;
                     let click_index = if (beat_index as u32) % self.nominator == 0 {0} else {1};
                     if offset < left.len() {
                         self.clicks.push(Click {sound_index: click_index, position: 0, start_index: offset})
